mps: test runner: adjust delays when waiting for replies
Previously, while `runner.py` was waiting for a response from the RTS,
it would send an update-display (`D`) command every second.  In the
nested-VM setup, updating the display is fairly expensive due to the I/O
involved.  If the MPS got behind for any reason, in order to catch up,
it would need to process not only the pending command but all the
update-display commands that were sent while the test runner was
waiting.  Sending an update-display command every second thus severely
exacerbated the slow performance of the nested-VM setup.

This commit changes `runner.py` to send the update-display command every
10 seconds instead of every second.  The test suite now behaves more
reliably in the nested-VM setup.