mps: makefile: run cryptol codegen only when REGENERATE_SOURCES is set
Copies of the generated source files are checked into this repo.
However, depending on the order in which files are created by git, the
generated code might be older than the Cryptol models, and `make` will
thus try to run the code generators.  This commit disables the code
generator rules unless the user sets `REGENERATE_SOURCES`.